+++ tests/unit/devices/BASE.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { GW2000 } from './../../../src/devices/GW2000';
+import { BASE } from './../../../src/devices/BASE';
 import { createPlatform, api } from './../../driver';
 
 //------------------------------------------------------------------------------
@@ -17,11 +17,15 @@ const dataReport = {
 const configs = ["v1Default", "v1Full", "v2Default", "v2Full"];
 
 configs.forEach(config => {
-  describe(`GW2000 device should be configured ${config}`, () => {
+  describe(`BASE device should be configured ${config}`, () => {
     before('Initialize platform', () => {
       platform = createPlatform(config);
       accessory = new api.platformAccessory('Accessory', "5746853e-4fee-4e47-97dd-53065ef1de03")
-      device = new GW2000(platform, accessory, "GW2000");
+
+      platform.config.nameOverrides = [];
+      platform.config.hidden = {};
+
+      device = new BASE(platform, accessory, "GW1000");
     });
 
     beforeEach('Reset config', () => {
@@ -46,10 +50,10 @@ configs.forEach(config => {
     });
 
     it('Services are created with name overrides', (done) => {
-      platform.config.nameOverrides[0] = {"key": "GW2000:indoorTemperature", "value": "Test Temperature Name"};
-      platform.config.nameOverrides[1] = {"key": "GW2000:indoorHumidity", "value": "Test Humidity Name"};
+      platform.config.nameOverrides[0] = {"key": "BASE:indoorTemperature", "value": "Test Temperature Name"};
+      platform.config.nameOverrides[1] = {"key": "BASE:indoorHumidity", "value": "Test Humidity Name"};
 
-      device = new GW2000(platform, accessory, "GW2000");
+      device = new BASE(platform, accessory, "GW1000");
 
       expect(device.temperature).to.not.be.undefined;
       expect(device.humidity).to.not.be.undefined;
@@ -62,7 +66,7 @@ configs.forEach(config => {
       platform.config.hidden["indoorTemperature"] = true;
       platform.config.hidden["indoorHumidity"] = true;
 
-      device = new GW2000(platform, accessory, "GW2000");
+      device = new BASE(platform, accessory, "GW1000");
 
       expect(device.temperature).to.be.undefined;
       expect(device.humidity).to.be.undefined;
@@ -70,10 +74,10 @@ configs.forEach(config => {
     });
 
     it('Services are not created when hidden with device-specific override', (done) => {
-      platform.config.hidden["GW2000:indoorTemperature"] = true;
-      platform.config.hidden["GW2000:indoorHumidity"] = true;
+      platform.config.hidden["BASE:indoorTemperature"] = true;
+      platform.config.hidden["BASE:indoorHumidity"] = true;
 
-      device = new GW2000(platform, accessory, "GW2000");
+      device = new BASE(platform, accessory, "GW1000");
 
       expect(device.temperature).to.be.undefined;
       expect(device.humidity).to.be.undefined;